+++ Frontend/src/App.jsx
@@ -7,6 +7,7 @@ import {
 import HomePage from './views/Home';
 import About from './views/About';
 import Imprint from './views/Imprint';
+import Terms from './views/Terms';
 import Privacy from './views/Privacy';
 import Docs from './views/Docs';
 import Contact from './views/Contact';
@@ -36,6 +37,7 @@ function App() {
           <Route path="/result" render={() => <VisualizationPage />} />
           <Route path="/explore" render={() => <Explore />} />
           <Route path="/imprint" render={() => <Imprint />} />
+          <Route path="/terms" render={() => <Terms />} />
           <Route path="/privacy" render={() => <Privacy />} />
 
         </Switch>
