+++ Frontend/src/components/Footer/index.jsx
@@ -86,6 +86,10 @@ export default function Footer() {
           <Typography fontSize="16px" color="black">Imprint</Typography>
         </LinkBox>
         <Typography fontSize="16px" sx={{ margin: '0 4px' }}> - </Typography>
+        <LinkBox to="/terms" style={{ textDecoration: "none" }}>
+          <Typography fontSize="16px" color="black">Terms of Usage</Typography>
+        </LinkBox>
+        <Typography fontSize="16px" sx={{ margin: '0 4px' }}> - </Typography>
         <LinkBox to="/privacy" style={{ textDecoration: "none" }}>
           <Typography fontSize="16px" color="black">Privacy Policy</Typography>
         </LinkBox>
